refactor(backup): provide commong implementation of interface
